+++ tests/test_versions.py
@@ -1,11 +1,7 @@
 import json
 from pathlib import Path
 import unittest
-
-try:
-    import tomllib
-except ModuleNotFoundError:  # Python 3.10
-    import tomli as tomllib
+import tomllib
 
 
 ROOT = Path(__file__).resolve().parent.parent